feat: merge script
