--- Assets/PersonalManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 718edc2c582b3b040941ca96ec9c6176
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
